major refactoring
