+++ potato/state.py
@@ -1,0 +1,13 @@
+"""
+Variables that keep track of user states.
+"""
+
+# A global mapping from username to the annotator's
+user_to_annotation_state = {}
+
+# A global mapping from an instance's id to its data.
+# This is filled by load_all_data()
+instance_id_to_data = {}
+
+# A global dict to keep tracking of the task assignment status
+task_assignment = {}
