+++ potato/user_config.json
@@ -1,2 +1,3 @@
 {"username": "123456@umich.edu", "email": "123456@umich.edu", "password": "8d969eef6ecad3c29a3a629280e686cf0c3f5d5a86aff3ca12020c923adc6c92"}
 {"username": "naitian@umich.edu", "email": "naitian@umich.edu", "password": "9f735e0df9a1ddc702bf0a1a7b83033f9f7153a00c29de82cedadc9957289b05"}
+{"username": "zxcv@zxcv.com", "email": "zxcv@zxcv.com", "password": "96cae35ce8a9b0244178bf28e4966c2ce1b8385723a96a6b838858cdd6ca0a1e"}
